benchmarking/requirements.txt
numpy>=1.26,<2.0.0
tqdm>=4.67,<5.0.0
pyyaml>=6.0.0,<7.0.0
h5py>=3.13.0,<4.0.0
